Add improved color-theme.js that fixes flashing
