+++ src/_webpack/preload.js
@@ -1,0 +1,8 @@
+// webpack entry point for JavaScript that is to be loaded before page render.
+
+// Copyright 2023 Qi Tianshi. All rights reserved.
+
+
+import ColorTheme from "./utils/color-theme";
+
+ColorTheme.initPreload();
